requirements.txt
asgiref==3.8.1
boto3==1.37.27
botocore==1.37.27
certifi==2025.1.31
dj-database-url==2.3.0
Django==5.1.7
django-cors-headers==4.7.0
django-filter==25.1
django-storages==1.14.6
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
gunicorn==23.0.0
jmespath==1.0.1
packaging==24.2
pillow==11.1.0
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-http-client==3.3.7
s3transfer==0.11.4
sendgrid==6.11.0
sentry-sdk==2.25.1
six==1.17.0
sqlparse==0.5.3
starkbank-ecdsa==2.2.0
typing_extensions==4.13.1
tzdata==2025.2
urllib3==2.3.0
whitenoise==6.9.0
